Made 'pi' an optional argument for dmw_calculation
'pi' is the (limiting) ratio of R/P; the size of the training
sample to the size of the test sample. I want to experiment with
setting it differently in my mixed_window statistic to see if
that improves monte carlo results, so now the mixed_window
functions set it to be Inf instead of the empirical values.